Moves clickstream golang images to multi-stage builds
The Alpine 3.7 repositories bumped go from 1.9.2 to 1.9.4, which caused our builds to break. Go 1.9.4 switches the way flags are passed to the linker, which causes problems for our go build commands. These images pull in librdkafka and produce a static build using -tags static, which no longer is passed to go correctly, causing the builds to fail. This commit builds moves building the binary to a first stage in golang:alpine (golang 1.10.0), and copies it into our standard alpine base image for distribution.