Replace cron with a shell script that sleeps
Cron _really_ doesn't like to run as any user other than root, and since
we only have a single job we want to run on an easily calculated
schedule we can replace it with a small shell script, then we don't have
to run our gc containers as root.

To make this work with the existing helm charts we catch "crond"
invocations in /entrypoint and run our script instead.