Pre-build python apks for slow compiling packages
The APKBUILD (and patches) were taken from the Alpine edge repos. The
only reason we need to rebuild them is so that they work with _our_
version of python (which is 3.7 as of alpine:3.10, Alpine edge is now on
3.8).

The exact commits/locations of the APKBUILD files included in this
commit:

- https://git.alpinelinux.org/aports/tree/main/cython/APKBUILD?id=646e3a59d7dfc38e3527e3f9e9518e8fc02d3661
- https://git.alpinelinux.org/aports/tree/main/py3-bcrypt/APKBUILD?id=a6733d86cf6e3446e496eb56dc1a661475098eb1
- https://git.alpinelinux.org/aports/tree/testing/py3-cassandra-driver/APKBUILD?id=7d989d745a09fd352495857f98c6b1d2518b9d22
- https://git.alpinelinux.org/aports/tree/main/py3-cryptography/APKBUILD?id=af3d5861ca9c2000eb6fdb664919b509e4e7c136
- https://git.alpinelinux.org/aports/tree/community/py3-pycryptodome/APKBUILD?id=dfd7dd095f1736998ef02119cda68ec77f123e71
- https://git.alpinelinux.org/aports/tree/testing/py3-fastavro/APKBUILD?id=06e72c68723c0637a9d856b4b869c179e584bf3a
- https://git.alpinelinux.org/aports/tree/testing/py3-grpcio/APKBUILD?id=0f73e4c7e79be62b20c4e5eb8eb7d3473ecd934b
- https://git.alpinelinux.org/aports/tree/community/py3-numpy/APKBUILD?id=1d6fbc11289f873ae1d9903b9d673a59eec7e92c
- https://git.alpinelinux.org/aports/tree/testing/py3-pandas/APKBUILD?id=43f79fbbe1a0e4975da7a2649c7a3214d7d0396a
- https://git.alpinelinux.org/aports/tree/community/py3-pycryptodome/APKBUILD?id=dfd7dd095f1736998ef02119cda68ec77f123e71
- https://git.alpinelinux.org/aports/tree/main/py3-pynacl/APKBUILD?id=4c4f0346a603a43b44ebf2622a0cd881fa0427c2
- https://git.alpinelinux.org/aports/tree/testing/py3-typed-ast/APKBUILD?id=c240d3f1ca8dfe9d6af5d6a736da39b9a58ba6aa

We don't store the private key anywhere, so each time we need to build
new packages we will generate a new keypair. (This is fine, as we don't
leave the image with our repo configured, so we don't ever push out
updates that we expect `apk update` to pick up.)