Changed prom port source label to pod annotation (#578)
We do not have a service for fluentd so we also do not have a service annotation for the scrape port. This was causing prometheus to scrape all container ports on fluentd which include a a couple istio ports that we don’t want scraped as part of the fluentd data.