Corrected rules that needed job names changed
Specifically this was hit by things that needed to have {{.Release.Name}}  rather than just kube-state with out the release name.

Removed the release namespace restrictions for when we are not in a singlenamespace

Locking things down to only the release namespace limits kube system components and node information so we need to collect that in the cloud. However in a single namespace enterprise setup we probably don’t want to poke around customers clusters so these extra metrics will not be available in that scenario.

Fixed alert annotations to they display correctly in slack based in our alertmanager config