Patches from 0.12 release (#617)
* Set Orbit and Houston tags to release-0.12

* Bump to 0.12.0-alpha.2

* Replicas pick (#583)

* Increment chart version

* Version as 0.12.0-alpha.1

* Configurable replicas for commander, houston

* Update to versioned images

* Bump versions

* Default airflow chart 0.12.0

* Update Houston to 0.12.1

* 0.12.0-alpha.4

* Release 0.12

* Rollback pick (#588)

* Increment chart version

* Version as 0.12.0-alpha.1

* Rollback nginx to 0.11.0

* Add more Prometheus alerting and use rules

* Increment chart version

* Version as 0.12.0-alpha.1

* Increment version

* Removed rules and added a ton of new rules

Removed redundant or non-optimal rules and added in a bunch of community standard rules and practices. Some of these however may not work yet because not all the metrics are being collected or job names may be different.

renamed recording rules

Added lots of curly braces because helm

changed job names of kube-state-metrics to kube-state

Added missing checks for persistence enabled.

Removed duplicate servierty property

fixed persistence logic

removed nginx down specific warning

This should be covered by the generic targetdown alerts.
Also changed urgent to critical

Updates to serverity levels

Added severity definitons

Moved around “platform” alerts that fit better in other groups.

Platform group is now specific to the astronomer platform and not the cloud infrascture

Changed description of scheduler alert based on feedback

Removed the persistence enable checks I added. Were not needed

removed kubelet down alert

This is not needed because we can’t scrape directly in gke, and we already are accomplishing the same goal by scapring kubelet metrics through the api metric proxy.

Changed prom port source label to pod annotation (#578)

We do not have a service for fluentd so we also do not have a service annotation for the scrape port. This was causing prometheus to scrape all container ports on fluentd which include a a couple istio ports that we don’t want scraped as part of the fluentd data.

Updated CPUThrottlingHigh rule to ignore all customer namespaces

it matches on namespaces that do not start with astronomer- and end with 4 digits

added tier: platform to all the new alerts

need this until new alert routing scheme is in place.

Update charts/prometheus/templates/prometheus-statefulset.yaml

Update charts/prometheus/templates/prometheus-statefulset.yaml

Update charts/prometheus/templates/prometheus-statefulset.yaml

Update charts/prometheus/templates/prometheus-statefulset.yaml

stateful set reverted back to original.

Revert statefulset change

Co-authored-by: James Weber <jpweber@gmail.com>

* Update font-src policy to allow Astronomer CDN (#591)

* Upgrade astro-ui to 0.12.1

* Updated kube-state-metrics chart to take replicas as a variable (#599)

Set the default value in the values file to be 1. Expecting us to override this in our cloud deployment.

This is work related for issue https://github.com/astronomer/issues/issues/991

* Corrected rules that needed job names changed (#598)

Specifically this was hit by things that needed to have {{.Release.Name}}  rather than just kube-state with out the release name.

Removed the release namespace restrictions for when we are not in a singlenamespace

Locking things down to only the release namespace limits kube system components and node information so we need to collect that in the cloud. However in a single namespace enterprise setup we probably don’t want to poke around customers clusters so these extra metrics will not be available in that scenario.

Fixed alert annotations to they display correctly in slack based in our alertmanager config

* Add additional domain to policy for Segment API (#596)

* Bump houston and UI

* Dropping things that start with id was causing duplicate records (#601)

With removing the restriction of astronomer namespace only it allows us to get node specific container metrics. the id label is part of this information, by dropping that label these were causing non-unique metrics because that id label was needed to keep them unique.

* Upgrade astro-ui, airflow chart versions

* Reconfigure commander liveness probe

* Upgrade Astro UI version (#610)

* Upgrade Astro UI version

* Increment chart version

Co-authored-by: sjmiller609 <sjmiller609@gmail.com>

* Update Chart.yaml

* Remove PR template

Co-authored-by: schnie <greg@astronomer.io>
Co-authored-by: James Weber <jpweber@gmail.com>
Co-authored-by: Ryan Hamilton <ryan@ryanahamilton.com>