Removed the init container added security context
The fields specified in the security context override the need to have an init container changing ownership.

This may have been needed in the past but the promethues processes runs as `nobody`  (65534) and `nogroup` (65534). The data is already owned by that user from the previous deployments because we keep the disk around.