Fix the Houston Worker Queries (#957)
The Houston Worker queries have a bug where the query and application code does not filter out deletedAt deployments from the NATS Message.  Normally this isn't a problem, but there is a race condition that we found while doing automated and rapid testing where we quickly create and delete deployments.  The NATS message can get into a bad state in this case and continue to resend messages that are unacknowledged.